👨‍💻 Kodlama yapısında köklü değişiklikler
- 🏗️ Proje dizin yapısı belirlendi
- 💦 Gereksizler kaldırıldı
- 👨‍🔬 Deneysel sürümde
